Change __error attribute to __error_if_used
Mac OS X seems to use __error in their system headers, and this
definition conflicts. __error_if_used is a more expressive name
anyway, so adjust this name.

I don't have a Mac system to test on, so this is just a change on the
presumption that it will improve the Mac situation for
compatibility. There might still be other issues with this code and
Mac OS X.